Update druid version. Address review comments
