+++ druid/src/contexts.rs
@@ -26,9 +26,9 @@ use crate::env::KeyLike;
 use crate::piet::{Piet, PietText, RenderContext};
 use crate::shell::Region;
 use crate::{
-    commands, Affine, Command, ContextMenu, Cursor, Env, ExtEventSink, Insets, MenuDesc,
-    Notification, Point, Rect, SingleUse, Size, SubWindowRequirement, Target, TimerToken, WidgetId, WindowDesc,
-    WindowHandle, WindowId,
+    commands, sub_window::SubWindowRequirement, widget::Widget, Affine, Command, ContextMenu,
+    Cursor, Data, Env, ExtEventSink, Insets, MenuDesc, Notification, Point, Rect, SingleUse, Size,
+    Target, TimerToken, WidgetId, WindowConfig, WindowDesc, WindowHandle, WindowId,
 };
 
 /// A macro for implementing methods on multiple contexts.
@@ -354,12 +354,17 @@ impl_context_method!(EventCtx<'_, '_>, UpdateCtx<'_, '_>, LifeCycleCtx<'_, '_>, 
     /// Create a new sub window that will have its app data synchronised with the nearest surrounding widget pod.
     // TODO - dynamically check that the type of the pod we are registering this on is the same as the type of the
     // requirement. Needs type ids recorded. This goes wrong if you don't have a pod between you and a lens.
-    pub fn new_sub_window(&mut self, requirement: SubWindowRequirement) {
-        if let Some(id) = requirement.host_id {
-            self.widget_state.add_sub_window_host(id);
-        }
-
-        self.submit_command(commands::NEW_SUB_WINDOW.with(SingleUse::new(requirement)));
+    pub fn new_sub_window<W: Widget<U> + 'static, U: Data>(
+        &mut self,
+        window_config: WindowConfig,
+        widget: W,
+        data: U,
+    ) -> WindowId {
+        let req = SubWindowRequirement::new(self.widget_id(), window_config, widget, data);
+        let window_id = req.window_id;
+        self.widget_state.add_sub_window_host(req.host_id);
+        self.submit_command(commands::NEW_SUB_WINDOW.with(SingleUse::new(req)));
+        window_id
     }
 });
